0.0.4
├─• Corrección de errores de web client de la api perfiles
├─• Corrección de test de Api perfiles
├─• Implementación de test de Api Registros
├─• Se corrige el dataloader de Api perfiles
├─• Se implementa la lógica y web client de la api de Incidentes
└─• Se inician las pruebas unitarias de incidentes

->>> Terminar pruebas unitarias de api de incidentes

+++ SAFE-Rescue_Backend_API-Incidentes/src/main/java/com/SAFE_Rescue/API_Incidentes/repository/IncidenteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface IncidenteRepository extends JpaRepository<Incidente, Long> {
+public interface IncidenteRepository extends JpaRepository<Incidente, Integer> {
 
 
 }
